Add `IdentifiedCollections` trait
