+++ Package@swift-5.9.swift
@@ -52,3 +52,10 @@ let package = Package(
     ),
   ]
 )
+
+for target in package.targets {
+  target.swiftSettings = target.swiftSettings ?? []
+  target.swiftSettings?.append(contentsOf: [
+    .define("IdentifiedCollections"),
+  ])
+}
