+++ Sources/Sharing/Internal/Reference.swift
@@ -1,6 +1,5 @@
 import Dependencies
 import Foundation
-import IdentifiedCollections
 import IssueReporting
 public import PerceptionCore
