+++ Sources/Sharing/Shared.swift
@@ -1,7 +1,6 @@
 public import CustomDump
 import Dependencies
 import Foundation
-import IdentifiedCollections
 public import Observation
 public import PerceptionCore
