+++ Sources/Sharing/SharedReader.swift
@@ -1,6 +1,5 @@
 public import CustomDump
 import Foundation
-import IdentifiedCollections
 import IssueReporting
 public import Observation
 public import PerceptionCore
