+++ Tests/SharingTests/SharedTests.swift
@@ -1,7 +1,6 @@
 import CustomDump
 import Dependencies
 import Foundation
-import IdentifiedCollections
 import PerceptionCore
 import Sharing
 import Testing
